requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
curl_cffi==0.11.4
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.58.5
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
idna==3.10
imbalanced-learn==0.13.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
multitasking==0.0.11
narwhals==1.47.0
numpy==2.3.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
patsy==1.0.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pyparsing==3.2.3
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scikit-learn==1.6.1
scipy==1.16.0
seaborn==0.13.2
selenium==4.34.2
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
statsmodels==0.14.5
streamlit==1.46.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yfinance==0.2.65
